misc: pmic8058-xoadc: Add mpp_config call on configure
Call mpp_config during xoadc read request. This is done
to configure the mpp mapping on scenario's when the remote
processor overwrites the configuration. The corner case for
this scenario is if a overwrite to the MPP occures in the
middle of the ADC transaction. By design, the  MPP mapping
is configured to be the same between the ADC drivers.

Change-Id: I7dfec84de0050f8e71be74b016c672eec0918091
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>