msm: restart: call secure code to disable arbiter
The other masters in the MSM could be accessing the pmic while
pmic has raised the reset interrupt. This might cause pmic
lockups.

Disable the pmic arbiter by calling secure code when we handle
the reset interrupt from pmic. This will disable all the ssbi
transactions to the pmic avoiding a possible lockup.

Also since we wont be lowering ps_hold for pmic initiated startups
piggy back on lower_pshold to return from the power off function
so that we can call in the secure manager as the last step.

Additionally, since we will be busy looping until the pmic resets
us, pet the watchdog.

CRs-Fixed: 306410
Change-Id: Iaf10c56b249c15f4ea0f43c7efb7412ff04004fb
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
heavily modified by faux123